Create main.cpp
This program increments or decrements each letter (alphabetical character) according to the enter key, which can be input as a parameter, using a terminal or setting it up on the IDE configurations. It works both for lowercase and uppercase characters. Special characters are displayed as question marks ('?') and spaces are not scrambled.